refactor: Extract car game class from App.js
